backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
websockets==12.0
python-multipart==0.0.6
pyserial==3.5
python-dotenv==1.0.0
typing-extensions==4.8.0 